AstraDB_Learnings/requirements.txt
#nstall the following packages 
astrapy
python-dotenv
langchain-astradb
langchain-community
langchain-huggingface
torch
transformers
sentence-transformers
